+++ app/Http/Controllers/Admin/BukuPanduanController.php
@@ -1,3 +1,7 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
 use App\Http\Controllers\Controller;
 use App\Models\BukuPanduan;
 use App\Models\KonsentrasiKeahlian;