+++ app/Http/Controllers/Siswa/PanduanController.php
@@ -4,9 +4,7 @@ namespace App\Http\Controllers\Siswa;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-
 use App\Models\BukuPanduan;
-use Illuminate\Http\Request;
 
 class PanduanController extends Controller
 {